+++ src/components/ContractStatusWidget/ContractStatus.js
@@ -1,0 +1,11 @@
+import { gql } from "@apollo/client";
+
+export const GET_CONTRACT_STATUS_COUNTS = gql`
+  query GetContractStatusCounts {
+    contractStatusCounts {
+      notStarted
+      inProgress
+      done
+    }
+  }
+`;
