--- src/context/apolloClient.js
@@ -1,12 +1,0 @@
-import { ApolloClient, InMemoryCache, HttpLink } from "@apollo/client";
-
-const httpLink = new HttpLink({
-  uri: "http://your-backend-uri/graphql",
-});
-
-const client = new ApolloClient({
-  link: httpLink,
-  cache: new InMemoryCache(),
-});
-
-export default client;
